requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
locust==2.20.1
huggingface_hub==0.20.3
prometheus_client==0.19.0
requests==2.31.0
python-dotenv==1.0.1
transformers==4.37.0
torch --index-url https://download.pytorch.org/whl/cpu
